src/requirements/base.txt
# Core
Django>=1.10,<1.11
Pillow>=3.1,<3.2
psycopg2>=2.6,<2.7

django-cors-headers>=1.1,<1.2
django-extensions>=1.6,<1.7
django-filter>=0.11,<0.12
django-rest-swagger>=0.3,<0.4
djangorestframework>=3.3,<3.4
djangorestframework-jwt>=1.7,<1.8
drf-nested-routers>=0.10,<0.11
djangorestframework-camel-case

# Solr
drf-haystack==1.6.1
django-haystack==2.5.1
pysolr==3.5.0

# Memcache
python-memcached==1.53

# Packages
ipython
Werkzeug
celery[librabbitmq]>=3.1,<3.2
